feat(ui): add collapsed version of widget
